requirements.txt
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
chromadb
sentence-transformers
pypdf
google-generativeai
streamlit
python-dotenv